backend/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-bcrypt
flask-cors
# Roboflow inference-sdk currently does not publish builds for Python 3.13.
inference-sdk; python_version < "3.13"
python-dotenv
pytest
